requirements.txt
google-cloud-core==2.4.1
joblib==1.4.0
jupyter==1.0.0
kfp==2.7.0
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
pydantic==2.7.0
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
google-cloud-pipeline-components==2.14.0